ISSUE #868 Refactor Proxy object related changes to seperate class
