+++ bouncer/src/repo/manipulator/modelconvertor/import/odaHelper/data_processor_dwg.h
@@ -28,20 +28,11 @@
 #include <vector>
 #include <unordered_map>
 #include <unordered_set>
-#include <array>
-#include <DbProxyEntity.h>
-#include <DbEntityWithGrData.h>
-#include <DbRegAppTable.h>
-#include <DbRegAppTableRecord.h>
-#include <DbDictionary.h>
-#include <DbXrecord.h>
 
 #include "vectorise_device_dgn.h"
 #include "repo/core/model/bson/repo_node_mesh.h"
 #include "repo/lib/datastructure/repo_variant.h"
-#include "proxy_app_handler.h"
-#include "civil3d_proxy_handler.h"
-#include "plant3d_proxy_handler.h"
+#include "dwg_proxy_inspector.h"
 
 namespace repo {
 	namespace manipulator {
@@ -54,11 +45,15 @@ namespace repo {
 					void setMode(OdGsView::RenderMode mode);
 					~DataProcessorDwg();
 
+					// The single DwgProxyInspector instance for the file being imported,
+					// shared by every DataProcessorDwg view instance ODA creates while
+					// vectorising it. Owned by FileProcessorDwg::importModel(), alongside
+					// the GeometryCollector; see dwg_proxy_inspector.h for why it needs
+					// to be file-scoped rather than per-instance.
+					void setProxy(DwgProxyInspector* proxy) { this->proxy = proxy; }
+
 					struct DiagnosticStats {
 						size_t totalEntities = 0;
-						size_t civil3dEntities = 0;
-						size_t plant3dEntities = 0;
-						size_t proxyEntities = 0;
 						size_t entitiesWithGeometry = 0;
 						size_t entitiesWithoutGeometry = 0;
 						std::unordered_map<std::string, size_t> entityTypeCount;
@@ -119,96 +114,6 @@ namespace repo {
 						const OdGiVertexData* pVertexData = 0) override;
 
 				private:
-					/* ===== Proxy entity inspection =====
-
-					A proxy entity is inspected once via getProxyInfo(), which performs the
-					single OdDbProxyEntity::cast() for that entity. The resulting ProxyInfo is
-					reused for classification, TIN detection, stored graphics replay, metadata
-					and diagnostics, instead of every consumer re-casting ODA.
-
-					The XData chain and the extension dictionary are deliberately not read up
-					front: opening the extension dictionary is a database object open, and most
-					proxies never need it - they produce no geometry, or their layer already
-					carries metadata. Those reads are loaded on demand by ensureProxyXData() and
-					ensureProxyExtensionDictionary(), which are no-ops once cached. */
-
-					struct ProxyReadOptions
-					{
-						/* Defaults are the cheap draw-time profile; metadata and
-						diagnostic paths opt in to what they actually need. */
-						bool readXData = false;
-						bool readExtensionDictionary = false;
-						/* Falls back to scanning the whole registered application
-						table when xData() yields nothing. Expensive, so off
-						outside diagnostics. */
-						bool scanRegisteredAppsFallback = false;
-					};
-
-					struct ProxyInfo
-					{
-						OdDbProxyEntityPtr entity; // null => not a proxy
-						std::string originalClass;
-						std::string originalDxfName;
-						std::string applicationDescription;
-						std::vector<std::string> xDataApps;
-						ProxyAppType appType = ProxyAppType::Unknown;
-						OdDbProxyEntity::GraphicsMetafileType graphicsType = OdDbProxyEntity::kNoMetafile;
-						bool hasFullGraphicsFlag = false;
-						OdDbEntityWithGrDataPEPtr graphicsPE;
-
-						/* Lazily populated; see ensureProxyXData() and
-						ensureProxyExtensionDictionary(). */
-						OdResBufPtr xData;
-						bool xDataLoaded = false;
-						OdDbObjectId extensionDictionaryId;
-						OdDbDictionaryPtr extensionDictionary;
-						bool extensionDictionaryLoaded = false;
-
-						// Set together with appType by classifyApplication(); never diverges from it.
-						ProxyAppHandler* matchedHandler = nullptr;
-
-						bool isProxy() const { return !entity.isNull(); }
-						bool isCivil3D() const { return appType == ProxyAppType::Civil3D; }
-						bool isPlant3D() const { return appType == ProxyAppType::Plant3D; }
-						bool hasFullGraphics() const { return hasFullGraphicsFlag; }
-					};
-
-					bool getProxyInfo(OdDbEntityPtr entity, ProxyInfo& info, const ProxyReadOptions& options = ProxyReadOptions());
-
-					/* On demand loaders for the two expensive proxy reads. Both return
-					immediately once the corresponding data has been cached in info. */
-					void ensureProxyXData(ProxyInfo& info, const ProxyReadOptions& options = ProxyReadOptions());
-					void ensureProxyExtensionDictionary(ProxyInfo& info);
-
-					ProxyAppType classifyApplication(const std::string& originalClass, ProxyAppHandler*& outHandler);
-					static std::string formatApplicationDisplayString(const ProxyInfo& info);
-
-					/* The class-name check formerly on ProxyInfo::isCivil3DTinSurface(),
-					moved here since a plain data struct can't reach civil3DHandler. */
-					bool isCivil3DTinSurface(const ProxyInfo& info) const
-					{
-						return info.isCivil3D() && civil3DHandler.isTinSurfaceClass(info.originalClass);
-					}
-
-					bool drawStoredProxyGraphics(OdDbEntityPtr pEntity, const ProxyInfo& info);
-					void logProxyWithoutRenderableGeometry(
-						OdDbEntityPtr pEntity,
-						ProxyInfo& info,
-						bool replayedStoredProxyGraphics,
-						bool replayReturnedGeometry);
-
-					std::unordered_map<std::string, repo::lib::RepoVariant> getProxyEntityMetadata(OdDbEntityPtr pEntity, ProxyInfo& info);
-					void addProxyBasicMetadata(OdDbEntityPtr pEntity, const ProxyInfo& info, std::unordered_map<std::string, repo::lib::RepoVariant>& metadata);
-					void addProxyGeneralMetadata(OdDbEntityPtr pEntity, std::unordered_map<std::string, repo::lib::RepoVariant>& metadata);
-					void addProxyGeometryMetadata(OdDbEntityPtr pEntity, std::unordered_map<std::string, repo::lib::RepoVariant>& metadata);
-					void addProxyXDataMetadata(const ProxyInfo& info, std::unordered_map<std::string, repo::lib::RepoVariant>& metadata);
-					void addProxyDictionaryMetadata(const ProxyInfo& info, std::unordered_map<std::string, repo::lib::RepoVariant>& metadata);
-
-					void extractXDataProperties(OdResBufPtr pRb, std::unordered_map<std::string, repo::lib::RepoVariant>& metadata);
-					void extractTextPropertiesFromProxy(OdDbProxyEntityPtr proxyEntity, std::unordered_map<std::string, repo::lib::RepoVariant>& metadata);
-					void extractEntityProperties(OdDbEntityPtr pEntity, std::unordered_map<std::string, repo::lib::RepoVariant>& metadata);
-					void removeDuplicateGeneralMetadata(
-						std::unordered_map<std::string, repo::lib::RepoVariant>& metadata);
 					void setEntityMetadata(
 						const std::string& layerId,
 						const std::string& handleMetaValue,
@@ -252,26 +157,8 @@ namespace repo {
 					Context context;
 					mutable DiagnosticStats stats;
 
-					/* Registered proxy-app handlers, tried in order for each proxy entity.
-					Both are always constructed (never conditionally) so a drawing containing
-					both Civil3D and Plant3D proxies classifies each entity independently.
-					Declared before proxyHandlers so taking their addresses below is
-					self-evidently safe regardless of member-initializer-list order. */
-					Civil3DProxyHandler civil3DHandler;
-					Plant3DProxyHandler plant3DHandler;
-					std::array<ProxyAppHandler*, 2> proxyHandlers = { &civil3DHandler, &plant3DHandler };
-
-					// Set for the duration of one entity's stored-graphics replay; the
-					// geometry callbacks route into it when non-null. Replaces the old
-					// TinSurfaceCapture tinCapture member/isActive() toggle.
-					ProxyGeometryCapture* activeGeometryCapture = nullptr;
-
-					/* Proxies that produced no renderable geometry are reported once each,
-					but only for the first kMaxLoggedProxyGeometryFailures of them; the rest
-					are counted and summarised, so a proxy heavy drawing cannot flood the log. */
-					static const size_t kMaxLoggedProxyGeometryFailures = 20;
-					std::unordered_set<std::string> loggedProxyGeometryFailures;
-					size_t suppressedProxyGeometryFailures = 0;
+					// Not owned; see setProxy().
+					DwgProxyInspector* proxy = nullptr;
 
 					std::string getClassDisplayName(OdDbEntityPtr entity, const ProxyInfo& info);
 				};
